h264: migrate to oxideav_core::bits + golomb extension trait
The Exp-Golomb codes, trailing-bits framing, and more_rbsp_data
sentinel are H.264-specific and stay in this crate as a new
src/golomb.rs module with BitReaderExt + BitWriterExt traits on
oxideav_core::bits::{BitReader, BitWriter}. The generic
bit reader + writer (src/bitreader.rs, src/bitwriter.rs, ~550 lines)
are deleted.

rbsp_to_ebsp (byte-layer emulation-prevention insertion) moves from
src/bitwriter.rs to src/nal.rs where it already belongs conceptually.

Call-site adjustments:
- write_bit(u32) → write_bit(bool) — the shared API's signature is
  bool so cavlc_enc.rs literal 0/1 become false/true and the
  sign_bit path gets `sign_bit != 0`.
- Files that call read_ue/se/te/flag or write_ue/se/te/flag +
  rbsp_trailing_bits now `use crate::golomb::{BitReaderExt,
  BitWriterExt};` alongside the oxideav_core::bits import.
- Redundant `br.read_se()? as i32` / `... as u32` casts collapse to
  `br.read_se()?` under clippy::unnecessary_cast.

126 lib tests + 2 integration tests green. The pre-existing clippy
debt on master (~80 warnings across unrelated files — lifetimes, doc
formatting, idiom tweaks) is not fixed here; out of scope for this
migration and predates the session.

+++ src/b_mb_444.rs
@@ -15,8 +15,8 @@
 
 use oxideav_core::{Error, Result};
 
-use crate::bitreader::BitReader;
 use crate::cavlc::{decode_residual_block, BlockKind};
+use crate::golomb::BitReaderExt;
 use crate::mb::LUMA_BLOCK_RASTER;
 use crate::mb_444::{self as mb444};
 use crate::mb_type::{
@@ -32,6 +32,7 @@ use crate::pps::Pps;
 use crate::slice::{LumaWeight, PredWeightTable, SliceHeader};
 use crate::sps::Sps;
 use crate::transform::{chroma_qp, dequantize_4x4_scaled, idct_4x4};
+use oxideav_core::bits::BitReader;
 
 use crate::b_mb::{
     apply_motion_compensation_b, direct_16x16_compensate_pub, direct_8x8_compensate_pub, BSliceCtx,
@@ -382,12 +383,12 @@ fn handle_16x16_single_dir(
         None
     };
     let mvd_l0 = if uses_l0(dir) {
-        Some((br.read_se()? as i32, br.read_se()? as i32))
+        Some((br.read_se()?, br.read_se()?))
     } else {
         None
     };
     let mvd_l1 = if uses_l1(dir) {
-        Some((br.read_se()? as i32, br.read_se()? as i32))
+        Some((br.read_se()?, br.read_se()?))
     } else {
         None
     };
@@ -428,12 +429,12 @@ fn handle_two_partition(
     }
     for p in 0..2 {
         if uses_l0(dirs[p]) {
-            mvd_l0[p] = Some((br.read_se()? as i32, br.read_se()? as i32));
+            mvd_l0[p] = Some((br.read_se()?, br.read_se()?));
         }
     }
     for p in 0..2 {
         if uses_l1(dirs[p]) {
-            mvd_l1[p] = Some((br.read_se()? as i32, br.read_se()? as i32));
+            mvd_l1[p] = Some((br.read_se()?, br.read_se()?));
         }
     }
     for p in 0..2 {
@@ -498,12 +499,12 @@ fn handle_b8x8(
         for _ in 0..n {
             if let Some(dir) = dir_opt {
                 let mvd_l0 = if uses_l0(dir) {
-                    Some((br.read_se()? as i32, br.read_se()? as i32))
+                    Some((br.read_se()?, br.read_se()?))
                 } else {
                     None
                 };
                 let mvd_l1 = if uses_l1(dir) {
-                    Some((br.read_se()? as i32, br.read_se()? as i32))
+                    Some((br.read_se()?, br.read_se()?))
                 } else {
                     None
                 };
